+++ puerh-jvm-executors/src/test/java/io/github/mishkun/puerh/handlers/SchedulingDecoratorSpec.kt
@@ -1,0 +1,9 @@
+package io.github.mishkun.puerh.handlers
+
+import io.kotest.core.spec.style.FreeSpec
+import io.kotest.experimental.robolectric.RobolectricTest
+
+@RobolectricTest
+class SchedulingDecoratorSpec : FreeSpec({
+    "describe"
+})
